requirements.txt
fastapi[all]
uvicorn[standard]
gunicorn
langchain
openai
requests
python-dotenv
python-decouple
pydantic
supabase
sse-starlette
pandas
tabulate
unstructured
